MAINT: stats: move entropy and differential_entropy to even more appropriate file
