+++ scipy/stats/__init__.py
@@ -444,6 +444,7 @@ from . import qmc
 from ._multivariate import *
 from . import contingency
 from .contingency import chi2_contingency
+from ._entropy import *
 
 __all__ = [s for s in dir() if not s.startswith("_")]  # Remove dunders.
